lib/fs: Consider win83 for root path as well when watching (ref #5706)
