+++ lib/fs/basicfs_watch.go
@@ -11,8 +11,6 @@ package fs
 import (
 	"context"
 	"errors"
-	"path/filepath"
-	"runtime"
 
 	"github.com/syncthing/notify"
 )
@@ -55,28 +53,6 @@ func (f *BasicFilesystem) Watch(name string, ignore Matcher, ctx context.Context
 	go f.watchLoop(name, root, backendChan, outChan, ignore, ctx)
 
 	return outChan, nil
-}
-
-// watchPaths adjust the folder root for use with the notify backend and the
-// corresponding absolute path to be passed to notify to watch name.
-func (f *BasicFilesystem) watchPaths(name string) (string, string, error) {
-	root, err := evalSymlinks(f.root)
-	if err != nil {
-		return "", "", err
-	}
-
-	// Remove `\\?\` prefix if the path is just a drive letter as a dirty
-	// fix for https://github.com/syncthing/syncthing/issues/5578
-	if runtime.GOOS == "windows" && filepath.Clean(name) == "." && len(root) <= 7 && len(root) > 4 && root[:4] == `\\?\` {
-		root = root[4:]
-	}
-
-	absName, err := rooted(name, root)
-	if err != nil {
-		return "", "", err
-	}
-
-	return filepath.Join(absName, "..."), root, nil
 }
 
 func (f *BasicFilesystem) watchLoop(name, evalRoot string, backendChan chan notify.EventInfo, outChan chan<- Event, ignore Matcher, ctx context.Context) {
