+++ lib/fs/basicfs_windows_test.go
@@ -44,6 +44,10 @@ func TestWindowsPaths(t *testing.T) {
 
 func TestResolveWindows83(t *testing.T) {
 	fs, dir := setup(t)
+	if isMaybeWin83(dir) {
+		dir = resolveWin83(dir)
+		fs = newBasicFilesystem(dir)
+	}
 	defer os.RemoveAll(dir)
 
 	shortAbs, _ := fs.rooted("LFDATA~1")
@@ -71,6 +75,10 @@ func TestResolveWindows83(t *testing.T) {
 
 func TestIsWindows83(t *testing.T) {
 	fs, dir := setup(t)
+	if isMaybeWin83(dir) {
+		dir = resolveWin83(dir)
+		fs = newBasicFilesystem(dir)
+	}
 	defer os.RemoveAll(dir)
 
 	tempTop, _ := fs.rooted(TempName("baz"))
